Refactor update

+++ sketches/src/test/java/com/yahoo/sketches/hll/HllMapTest.java
@@ -35,7 +35,7 @@ public class HllMapTest {
     key = Util.intToBytes(1, key);
     for (int i=1; i<= u; i++) {
       id = Util.intToBytes(i, id);
-      int coupon = Map.coupon16(id);
+      short coupon = (short) Map.coupon16(id);
       est = map.update(key, coupon);
       if (i % 100 == 0) {
         double err = (est/i -1.0) * 100;
@@ -77,7 +77,7 @@ public class HllMapTest {
       key = Util.intToBytes(j, key);
       for (i=0; i< u; i++) {
         id = Util.longToBytes(++v, id);
-        int coupon = Map.coupon16(id);
+        short coupon = (short) Map.coupon16(id);
         map.update(key, coupon);
       }
       double est = map.getEstimate(key);
